refactor(capture): split source-backed helpers
